--- a/src/monalisten/__init__.py
+++ b/src/monalisten/__init__.py
@@ -1,3 +1,8 @@
+from githubkit import lazy_module as lm
+
+lm.LAZY_MODULES = (*lm.LAZY_MODULES, r"^monalisten\.types$")
+lm.apply()
+
 from . import types
 from ._core import Monalisten