This plugs into `githubkit`'s lazy module mechanism by including `monalisten.types` in `LAZY_MODULES`. This is pretty hacky, but *it should be fine*. 🥲

Thanks to this, I was able to achieve the following speedups (on my M2 MBA):
* `uv run pytest` went down from ~4.47s to ~476ms (~9.4x speedup)
* importing `monalisten.types.PushEvent` went down from ~3.65s to ~156ms (~23.5x speedup)
* importing `monalisten.types.PullRequestEvent` went down ~3.65s to ~1.33s (~2.74x speedup)